chore: Wallet에서 MoneyUnitItem 컴포넌트로 분리
